+++ test/ziggurat/messaging/producer_test.clj
@@ -56,6 +56,26 @@
         (let [message-from-mq (rmq/get-msg-from-channel-dead-queue topic-entity channel)]
           (is (= expected-message-payload message-from-mq))))))
 
+  (testing "message in channel will be retried in delay queue with suffix 1 if message retry-count exceeds retry count in channel config"
+    (with-redefs [ziggurat-config (constantly (assoc (ziggurat-config)
+                                                     :stream-router
+                                                     {:default
+                                                      {:channels
+                                                       {:exponential-retry
+                                                        {:retry {:count 5
+                                                                 :enabled          true
+                                                                 :type             :exponential
+                                                                 :queue-timeout-ms 1000}}}}}))]
+      (fix/with-queues
+        {:default {:handler-fn #(constantly nil)
+                   :exponential-retry  #(constantly nil)}}
+        (let [channel                  :exponential-retry
+              retry-message-payload    (assoc message-payload :retry-count 10)
+              expected-message-payload (assoc message-payload :retry-count 9)
+              _                        (producer/retry-for-channel retry-message-payload channel)
+              message-from-mq          (rmq/get-message-from-channel-retry-queue topic-entity channel 1)]
+          (is (= expected-message-payload message-from-mq))))))
+
   (testing "message in channel will be retried with linear queue timeout"
     (with-redefs [ziggurat-config (constantly (assoc (ziggurat-config)
                                                      :stream-router {:default {:channels {:linear-retry {:retry {:count            5
@@ -77,13 +97,12 @@
           (let [message-from-mq (rmq/get-msg-from-channel-dead-queue topic-entity channel)]
             (is (= expected-message-payload message-from-mq)))))))
 
-  (testing "message in channel will be retried with exponential timeout calculated fro channel specific queue-timeout-ms value"
+  (testing "message in channel will be retried with exponential timeout calculated from channel specific queue-timeout-ms value"
     (with-redefs [ziggurat-config (constantly (assoc (ziggurat-config)
                                                      :stream-router {:default {:channels {:exponential-retry {:retry {:count               5
                                                                                                                       :enabled             true
-                                                                                                                      :queue-timeout-ms    1000
-                                                                                                                      :exponential-backoff {:enabled true
-                                                                                                                                            :count   10}}}}}}))]
+                                                                                                                      :type                :exponential
+                                                                                                                      :queue-timeout-ms    1000}}}}}))]
       (fix/with-queues
         {:default {:handler-fn        #(constantly nil)
                    :exponential-retry #(constantly nil)}}
@@ -92,28 +111,6 @@
               channel-retry-count      (:count (channel-retry-config topic-entity channel))
               retry-message-payload    (assoc message-payload :retry-count @retry-count)
               expected-message-payload (assoc message-payload :retry-count channel-retry-count)]
-          (producer/retry-for-channel retry-message-payload channel)
-          (while (> @retry-count 0)
-            (swap! retry-count dec)
-            (let [message-from-mq (rmq/get-message-from-channel-retry-queue topic-entity channel (- 5 @retry-count))]
-              (producer/retry-for-channel message-from-mq channel)))
-          (let [message-from-mq (rmq/get-msg-from-channel-dead-queue topic-entity channel)]
-            (is (= expected-message-payload message-from-mq)))))))
-
-  (testing "message in channel will be retried with exponential timeout calculated from queue-timeout-ms value"
-    (with-redefs [ziggurat-config (constantly (assoc (ziggurat-config)
-                                                     :stream-router {:default {:channels {:channel-exponential-retry {:retry {:count               5
-                                                                                                                              :enabled             true
-                                                                                                                              :exponential-backoff {:enabled true
-                                                                                                                                                    :count   10}}}}}}))]
-      (fix/with-queues
-        {:default {:handler-fn                #(constantly nil)
-                   :channel-exponential-retry #(constantly nil)}}
-        (let [retry-count              (atom 5)
-              channel                  :channel-exponential-retry
-              channel-retry-count      (:count (channel-retry-config topic-entity channel))
-              retry-message-payload    (assoc message-payload :retry-count @retry-count)
-              expected-message-payload (assoc message-payload :retry-count @retry-count)]
           (producer/retry-for-channel retry-message-payload channel)
           (while (> @retry-count 0)
             (swap! retry-count dec)
@@ -208,13 +205,30 @@
           (let [message-from-mq (rmq/get-msg-from-delay-queue "default")]
             (producer/retry message-from-mq)))
         (let [message-from-mq (rmq/get-msg-from-dead-queue "default")]
-          (is (= expected-message-payload message-from-mq)))))))
+          (is (= expected-message-payload message-from-mq))))))
+
+  (testing "[Backward Compatiblity] Messages will be retried even if retry type is not present in the config."
+    (with-redefs [ziggurat-config (constantly (update-in (ziggurat-config) [:retry] dissoc :type))]
+      (fix/with-queues
+        {:default {:handler-fn #(constantly nil)}}
+        (let [retry-count              (atom 2)
+              retry-message-payload    (assoc message-payload :retry-count @retry-count)
+              expected-message-payload (assoc message-payload :retry-count (retry-count-config))]
+          (producer/retry retry-message-payload)
+          (while (> @retry-count 0)
+            (swap! retry-count dec)
+            (let [message-from-mq (rmq/get-msg-from-delay-queue "default")]
+              (producer/retry message-from-mq)))
+          (let [message-from-mq (rmq/get-msg-from-dead-queue "default")]
+            (is (= expected-message-payload message-from-mq))))))))
 
 (deftest retry-with-exponential-backoff-test
   (testing "message will publish to delay with retry count queue when exponential backoff enabled"
     (with-redefs [ziggurat-config (constantly (assoc (ziggurat-config)
-                                                     :retry {:count 5 :enabled true :exponential-backoff {:enabled true :count 10}}))]
-      (testing "message with no retry count will publish to delay with retry count queue"
+                                                     :retry {:count 5
+                                                             :enabled true
+                                                             :type :exponential}))]
+      (testing "message with no retry count will publish to delay queue with suffix 1"
         (fix/with-queues
           {:default {:handler-fn #(constantly nil)}}
           (let [expected-message (assoc message-payload :retry-count 4)]
@@ -238,13 +252,23 @@
                 expected-message      (assoc message-payload :retry-count 0)]
             (producer/retry retry-message-payload)
             (let [message-from-mq (rmq/get-message-from-retry-queue "default" 5)]
-              (is (= message-from-mq expected-message)))))))))
+              (is (= message-from-mq expected-message))))))
+
+      (testing "message will be retried in delay queue with suffix 1 if message retry-count exceeds retry count in config"
+        (fix/with-queues
+          {:default {:handler-fn #(constantly nil)}}
+          (let [retry-message-payload    (assoc message-payload :retry-count 10)
+                expected-message-payload (assoc message-payload :retry-count 9)]
+            (producer/retry retry-message-payload)
+            (let [message-from-mq (rmq/get-message-from-retry-queue "default" 1)]
+              (is (= message-from-mq expected-message-payload)))))))))
 
 (deftest make-queues-test
   (let [ziggurat-config (ziggurat-config)]
     (testing "When retries are enabled"
-      (with-redefs [ziggurat-config (constantly (assoc ziggurat-config
-                                                       :retry {:enabled true}))]
+      (with-redefs [config/ziggurat-config (constantly (assoc ziggurat-config
+                                                              :retry {:enabled true
+                                                                      :type :linear}))]
         (testing "it does not create queues when stream-routes are not passed"
           (let [counter (atom 0)]
             (with-redefs [producer/create-and-bind-queue (fn
@@ -308,7 +332,100 @@
               (lq/delete ch dead-queue-name)
               (le/delete ch delay-exchange-name)
               (le/delete ch instant-exchange-name)
-              (le/delete ch dead-exchange-name))))))
+              (le/delete ch dead-exchange-name))))
+        (testing "it creates queues with suffixes in the range [1, retry-count] when exponential backoff is enabled"
+          (with-open [ch (lch/open connection)]
+            (let [stream-routes         {:default {:handler-fn #(constantly :success)}}
+                  retry-count           (get-in ziggurat-config [:retry :count])
+                  instant-queue-name    (util/prefixed-queue-name "default" (:queue-name (:instant (rabbitmq-config))))
+                  instant-exchange-name (util/prefixed-queue-name "default" (:exchange-name (:instant (rabbitmq-config))))
+                  delay-queue-name      (util/prefixed-queue-name "default" (:queue-name (:delay (rabbitmq-config))))
+                  delay-exchange-name   (util/prefixed-queue-name "default" (:exchange-name (:delay (rabbitmq-config))))
+                  dead-queue-name       (util/prefixed-queue-name "default" (:queue-name (:dead-letter (rabbitmq-config))))
+                  dead-exchange-name    (util/prefixed-queue-name "default" (:exchange-name (:dead-letter (rabbitmq-config))))
+                  expected-queue-status {:message-count 0 :consumer-count 0}
+                  exponential-delay-queue-name #(util/prefixed-queue-name delay-queue-name %)
+                  exponential-delay-exchange-name #(util/prefixed-queue-name delay-exchange-name %)]
+
+              (with-redefs [config/ziggurat-config (constantly (assoc-in ziggurat-config [:retry :type] :exponential))]
+                (producer/make-queues stream-routes)
+
+                (is (= expected-queue-status (lq/status ch dead-queue-name)))
+                (is (= expected-queue-status (lq/status ch instant-queue-name)))
+                (lq/delete ch instant-queue-name)
+                (lq/delete ch dead-queue-name)
+                (le/delete ch instant-exchange-name)
+                (le/delete ch dead-exchange-name)
+
+                ;; Verifying that delay queues with appropriate suffixes have been created
+                (doseq [s (range 1 (inc retry-count))]
+                  (is (= expected-queue-status (lq/status ch (exponential-delay-queue-name s))))
+                  (lq/delete ch (exponential-delay-queue-name s))
+                  (le/delete ch (exponential-delay-exchange-name s)))))))
+        (testing "it creates queues with suffixes in the range [1, 25] when exponential backoff is enabled and retry-count is more than 25"
+          (with-open [ch (lch/open connection)]
+            (let [stream-routes         {:default {:handler-fn #(constantly :success)}}
+                  instant-queue-name    (util/prefixed-queue-name "default" (:queue-name (:instant (rabbitmq-config))))
+                  instant-exchange-name (util/prefixed-queue-name "default" (:exchange-name (:instant (rabbitmq-config))))
+                  delay-queue-name      (util/prefixed-queue-name "default" (:queue-name (:delay (rabbitmq-config))))
+                  delay-exchange-name   (util/prefixed-queue-name "default" (:exchange-name (:delay (rabbitmq-config))))
+                  dead-queue-name       (util/prefixed-queue-name "default" (:queue-name (:dead-letter (rabbitmq-config))))
+                  dead-exchange-name    (util/prefixed-queue-name "default" (:exchange-name (:dead-letter (rabbitmq-config))))
+                  expected-queue-status {:message-count 0 :consumer-count 0}
+                  exponential-delay-queue-name #(util/prefixed-queue-name delay-queue-name %)
+                  exponential-delay-exchange-name #(util/prefixed-queue-name delay-exchange-name %)]
+
+              (with-redefs [config/ziggurat-config (constantly (-> ziggurat-config
+                                                                   (assoc-in [:retry :type] :exponential)
+                                                                   (assoc-in [:retry :count] 50)))]
+                (producer/make-queues stream-routes)
+                (is (= expected-queue-status (lq/status ch dead-queue-name)))
+                (is (= expected-queue-status (lq/status ch instant-queue-name)))
+                (lq/delete ch instant-queue-name)
+                (lq/delete ch dead-queue-name)
+                (le/delete ch instant-exchange-name)
+                (le/delete ch dead-exchange-name)
+                ;; Verifying that delay queues with appropriate suffixes have been created
+                (doseq [s (range 1 25)]
+                  (is (= expected-queue-status (lq/status ch (exponential-delay-queue-name s))))
+                  (lq/delete ch (exponential-delay-queue-name s))
+                  (le/delete ch (exponential-delay-exchange-name s)))))))
+        (testing "it creates delay queue for linear retries when retry type is not defined in the config"
+          (let [make-delay-queue-called (atom false)
+                stream-routes          {:default  {:handler-fn #(constantly nil)}}]
+            (with-redefs [config/ziggurat-config (constantly (update-in ziggurat-config [:retry] dissoc :type))
+                          producer/make-queue       (constantly nil)
+                          producer/make-delay-queue (fn [topic]
+                                                      (if (= topic :test)
+                                                        (reset! make-delay-queue-called true)))]
+              (producer/make-queues stream-routes))))
+        (testing "it creates delay queue for linear retries when retry type is incorrectly defined in the config"
+          (let [make-delay-queue-called (atom false)
+                stream-routes          {:default  {:handler-fn #(constantly nil)}}]
+            (with-redefs [config/ziggurat-config (constantly (assoc-in ziggurat-config [:retry :type] :incorrect))
+                          producer/make-queue       (constantly nil)
+                          producer/make-delay-queue (fn [topic]
+                                                      (if (= topic :test)
+                                                        (reset! make-delay-queue-called true)))]
+              (producer/make-queues stream-routes))))
+        (testing "it creates channel delay queue for linear retries when retry type is not defined in the channel config"
+          (let [make-channel-delay-queue-called (atom false)
+                stream-routes          {:default {:handler-fn #(constantly nil) :channel-1  {:handler-fn #(constantly nil)}}}]
+            (with-redefs [config/ziggurat-config    (constantly (update-in ziggurat-config [:stream-router :default :channels :channel-1 :retry] dissoc :type))
+                          producer/make-channel-queue       (constantly nil)
+                          producer/make-channel-delay-queue (fn [topic channel]
+                                                              (if (and (= channel :channel-1) (= topic :default))
+                                                                (reset! make-channel-delay-queue-called true)))]
+              (producer/make-queues stream-routes))))
+        (testing "it creates channel delay queue for linear retries when an incorrect retry type is defined in the channel config"
+          (let [make-channel-delay-queue-called (atom false)
+                stream-routes          {:default {:handler-fn #(constantly nil) :channel-1  {:handler-fn #(constantly nil)}}}]
+            (with-redefs [config/ziggurat-config    (constantly (assoc-in ziggurat-config [:stream-router :default :channels :channel-1 :retry :type] :incorrect))
+                          producer/make-channel-queue       (constantly nil)
+                          producer/make-channel-delay-queue (fn [topic channel]
+                                                              (if (and (= channel :channel-1) (= topic :default))
+                                                                (reset! make-channel-delay-queue-called true)))]
+              (producer/make-queues stream-routes))))))
 
     (testing "when retries are disabled"
       (with-redefs [config/ziggurat-config (constantly (assoc ziggurat-config
@@ -431,33 +548,36 @@
                                                                                                                           :enabled          true
                                                                                                                           :queue-timeout-ms 2000}}}}}))]
           (is (= 2000 (producer/get-channel-queue-timeout-ms topic-entity channel message))))))
-    (testing "when exponential backoff are enabled and channel retry count not defined"
-      (let [channel :channel-no-retry-count]
-        (with-redefs [config/ziggurat-config (constantly (assoc (config/ziggurat-config)
-                                                                :stream-router {topic-entity {:channels {channel {:retry {:enabled             true
-                                                                                                                          :exponential-backoff {:enabled true :count 10}}}}}}))]
-          (is (= 700 (producer/get-channel-queue-timeout-ms topic-entity channel message))))))
     (testing "when exponential backoff are enabled and channel queue timeout defined"
       (let [channel :exponential-retry]
         (with-redefs [config/ziggurat-config (constantly (assoc (config/ziggurat-config)
                                                                 :stream-router {topic-entity {:channels {channel {:retry {:count               5
                                                                                                                           :enabled             true
-                                                                                                                          :queue-timeout-ms    1000
-                                                                                                                          :exponential-backoff {:enabled true :count 10}}}}}}))]
+                                                                                                                          :type                :exponential
+                                                                                                                          :queue-timeout-ms    1000}}}}}))]
           (is (= 7000 (producer/get-channel-queue-timeout-ms topic-entity channel message))))))
-    (testing "when exponential backoff are enabled and channel queue timeout not defined"
-      (let [channel :channel-exponential-retry]
+
+    (testing "when exponential backoff are enabled and channel queue timeout is not defined"
+      (let [channel :exponential-retry]
         (with-redefs [config/ziggurat-config (constantly (assoc (config/ziggurat-config)
                                                                 :stream-router {topic-entity {:channels {channel {:retry {:count               5
                                                                                                                           :enabled             true
-                                                                                                                          :exponential-backoff {:enabled true}}}}}}))]
+                                                                                                                          :type                :exponential}}}}}))]
           (is (= 700 (producer/get-channel-queue-timeout-ms topic-entity channel message))))))))
 
 (deftest get-queue-timeout-ms-test
-  (testing "when retries are enabled"
+  (testing "when exponential retries are enabled"
     (let [message (assoc message-payload :retry-count 2)]
       (with-redefs [config/ziggurat-config (constantly (assoc (config/ziggurat-config)
                                                               :retry {:enabled             true
                                                                       :count               5
-                                                                      :exponential-backoff {:enabled true :count 10}}))]
-        (is (= 700 (producer/get-queue-timeout-ms message)))))))
+                                                                      :type                :exponential}))]
+        (is (= 700 (producer/get-queue-timeout-ms message))))))
+  (testing "when exponential retries are enabled and retry-count exceeds 25, the max possible timeouts are calculated using 25 as the retry-count"
+    (let [message (assoc message-payload :retry-count 20)]
+      (with-redefs [config/ziggurat-config (constantly (assoc (config/ziggurat-config)
+                                                              :retry {:enabled             true
+                                                                      :count               50
+                                                                      :type                :exponential}))]
+        ;; For 25 max exponential retries, exponent comes to 25-20=5, which makes timeout = 100*(2^5-1) = 3100
+        (is (= 3100 (producer/get-queue-timeout-ms message)))))))